fix as review
