+++ paddle/phi/kernels/yolov3_loss_grad_kernel.h
@@ -23,7 +23,7 @@ void Yolov3LossGradKernel(const Context& dev_ctx,
                           const DenseTensor& x,
                           const DenseTensor& gt_box,
                           const DenseTensor& gt_label,
-                          const paddle::optional<const DenseTensor&> gt_score,
+                          paddle::optional<const DenseTensor&> gt_score,
                           const DenseTensor& loss_grad,
                           const DenseTensor& objectness_mask,
                           const DenseTensor& gt_match_mask,
