+++ paddle/phi/kernels/yolov3_loss_kernel.h
@@ -23,7 +23,7 @@ void Yolov3LossKernel(const Context& dev_ctx,
                       const DenseTensor& x,
                       const DenseTensor& gt_box,
                       const DenseTensor& gt_label,
-                      const paddle::optional<const DenseTensor&> gt_score,
+                      paddle::optional<const DenseTensor&> gt_score,
                       const std::vector<int>& anchors,
                       const std::vector<int>& anchor_mask,
                       int class_num,
